Add extra padding bottom to footer discussion section & move link constant to lib
