+++ components/ui/footer-discussion-section.tsx
@@ -1,10 +1,10 @@
-import { Container } from "../ui/container";
-import { DISCUSSION_URL } from "~/constants/link";
+import { Container } from "~/components/ui/container";
+import { DISCUSSION_URL } from "~/lib/constants/link";
 
 export function FooterDiscussionSection() {
   return (
     <Container>
-      <section className="bg-white overflow-hidden p-4 mt-2">
+      <section className="bg-white overflow-hidden px-4 mt-2 pt-8 pb-16">
         <span>Ada usulan atau laporan terkait website ini? </span>
         <a
           className="underline text-blue-800"
